third_party/blink/web_tests/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-vertical-writing-mode-navigation.optional-expected.txt

This is a testharness.js-based test.
[FAIL] vertical-lr: Right/Left keys navigate as primary axis
  assert_equals: ArrowRight should move to the next option. expected Element node <option class="two">two</option> but got Element node <select id="vertical-lr" class="vertical-lr">\n  <option c...
[FAIL] vertical-rl: Left/Right keys navigate as primary axis (block-flipped)
  assert_equals: ArrowLeft should move to the next option. expected Element node <option class="two">two</option> but got Element node <select id="vertical-rl" class="vertical-rl">\n  <option c...
[FAIL] sideways-lr: Right/Left keys navigate as primary axis
  assert_equals: ArrowRight should move to the next option. expected Element node <option class="two">two</option> but got Element node <select id="sideways-lr" class="sideways-lr">\n  <option c...
[FAIL] sideways-rl: Left/Right keys navigate as primary axis (block-flipped)
  assert_equals: ArrowLeft should move to the next option. expected Element node <option class="two">two</option> but got Element node <select id="sideways-rl" class="sideways-rl">\n  <option c...
Harness: the test ran to completion.
